+++ single_kernel_mongo/core/operator.py
@@ -48,12 +48,12 @@ from single_kernel_mongo.workload.mongos_workload import MongosWorkload
 
 if TYPE_CHECKING:
     from single_kernel_mongo.abstract_charm import AbstractMongoCharm
-    from single_kernel_mongo.core.abstract_upgrades_v3 import MongoDBRefresh
     from single_kernel_mongo.events.database import DatabaseEventsHandler
     from single_kernel_mongo.events.tls import TLSEventsHandler
     from single_kernel_mongo.managers.ldap import LDAPManager
     from single_kernel_mongo.managers.tls import TLSManager
     from single_kernel_mongo.managers.upgrade_v3 import MongoDBUpgradesManager
+    from single_kernel_mongo.managers.upgrade_v3_status import MongoDBUpgradesStatusManager
 
 logger = getLogger(__name__)
 
@@ -80,10 +80,10 @@ class OperatorProtocol(ABC, Object, ManagerStatusProtocol):
     config_manager: FileBasedConfigManager
     tls_manager: TLSManager
     state: CharmState
-    upgrade_backend: MongoDBRefresh
     refresh: charm_refresh.Common | None
     mongo_manager: MongoManager
     upgrades_manager: MongoDBUpgradesManager
+    upgrades_status_manager: MongoDBUpgradesStatusManager
     ldap_manager: LDAPManager
     workload: MainWorkloadType
     client_events: DatabaseEventsHandler
